refactor: hardcode minecraft runtime into KairoRuntime class
