+++ lib/index.js
@@ -10902,14 +10902,7 @@ ${contextFunctionCode}`,
 });
 
 // src/index.ts
-import { world as world3 } from "@minecraft/server";
-
-// src/minecraft/MinecraftRuntime.ts
-import {
-  ScriptEventSource,
-  system,
-  world as world2
-} from "@minecraft/server";
+import { world as world2 } from "@minecraft/server";
 
 // src/utils/SeedRandom.ts
 var import_seedrandom = __toESM(require_seedrandom2(), 1);
@@ -10921,125 +10914,6 @@ var SeedRandom = class {
   next() {
     return this.rng();
   }
-};
-
-// src/minecraft/ScoreboardIdRegistry.ts
-import { world } from "@minecraft/server";
-var ScoreboardIdRegistry = class {
-  constructor(objectiveId) {
-    this.objectiveId = objectiveId;
-  }
-  get objective() {
-    const obj = world.scoreboard.getObjective(this.objectiveId);
-    if (!obj) {
-      throw new Error(`Objective not found: ${this.objectiveId}`);
-    }
-    return obj;
-  }
-  has(id) {
-    return this.objective.hasParticipant(id);
-  }
-  register(id) {
-    this.objective.setScore(id, 0);
-  }
-};
-
-// src/minecraft/minecraftEventBinding.ts
-var minecraftEventBinding = {
-  after: {
-    addonActivate: (_world, _handler) => {
-      return {
-        dispose: () => {
-        }
-      };
-    },
-    playerJoin: (world4, handler) => {
-      return world4.afterEvents.playerJoin.subscribe(handler);
-    }
-  },
-  before: {
-    addonDeactivate: (_world, _handler) => {
-      return {
-        dispose: () => {
-        }
-      };
-    }
-  }
-};
-
-// src/minecraft/MinecraftRuntime.ts
-var MinecraftRuntime = class {
-  constructor(options = {}) {
-    this.options = options;
-  }
-  currentTick() {
-    return system.currentTick;
-  }
-  send(id, message) {
-    system.sendScriptEvent(id, message);
-  }
-  receive(handler) {
-    const listener = (ev) => {
-      if (ev.sourceType !== ScriptEventSource.Server) return;
-      handler(ev.id, ev.message);
-    };
-    system.afterEvents.scriptEventReceive.subscribe(listener);
-    return {
-      dispose: () => {
-        system.afterEvents.scriptEventReceive.unsubscribe(listener);
-      }
-    };
-  }
-  onReady(handler) {
-    const listener = (_ev) => {
-      handler();
-    };
-    world2.afterEvents.worldLoad.subscribe(listener);
-    return {
-      dispose: () => {
-        world2.afterEvents.worldLoad.unsubscribe(listener);
-      }
-    };
-  }
-  createIdRegistry(objectiveId) {
-    return new ScoreboardIdRegistry(objectiveId);
-  }
-  createRandom() {
-    return new SeedRandom(this.options.randomSeed);
-  }
-  bindEvents(handler) {
-    const disposables = [];
-    for (const [name, fn] of Object.entries(minecraftEventBinding.after)) {
-      disposables.push(
-        fn(world2, (payload) => {
-          handler({
-            phase: "after",
-            name,
-            payload
-          });
-        })
-      );
-    }
-    for (const [name, fn] of Object.entries(minecraftEventBinding.before)) {
-      disposables.push(
-        fn(world2, (payload) => {
-          handler({
-            phase: "before",
-            name,
-            payload
-          });
-        })
-      );
-    }
-    return {
-      dispose: () => disposables.forEach((d) => d.dispose())
-    };
-  }
-  scheduler = {
-    runInterval: (cb, tick) => system.runInterval(cb, tick),
-    runTimeout: (cb, tick) => system.runTimeout(cb, tick),
-    clearRun: (id) => system.clearRun(id)
-  };
 };
 
 // src/router/events/classes/AddonActivateAfterEvent.ts
@@ -14810,6 +14684,101 @@ var ReadyState = class {
   }
 };
 
+// src/router/types/KairoRuntime.ts
+import {
+  ScriptEventSource,
+  system,
+  world
+} from "@minecraft/server";
+var ScoreboardIdRegistry = class {
+  constructor(objectiveId) {
+    this.objectiveId = objectiveId;
+  }
+  get objective() {
+    const obj = world.scoreboard.getObjective(this.objectiveId);
+    if (!obj) {
+      throw new Error(`Objective not found: ${this.objectiveId}`);
+    }
+    return obj;
+  }
+  has(id) {
+    return this.objective.hasParticipant(id);
+  }
+  register(id) {
+    this.objective.setScore(id, 0);
+  }
+};
+var minecraftEventBinding = {
+  after: {
+    addonActivate: (_world, _handler) => ({ dispose: () => {
+    } }),
+    playerJoin: (world3, handler) => world3.afterEvents.playerJoin.subscribe(handler)
+  },
+  before: {
+    addonDeactivate: (_world, _handler) => ({ dispose: () => {
+    } })
+  }
+};
+var KairoRuntime = class {
+  constructor(options = {}) {
+    this.options = options;
+  }
+  currentTick() {
+    return system.currentTick;
+  }
+  send(id, message) {
+    system.sendScriptEvent(id, message);
+  }
+  receive(handler) {
+    const listener = (ev) => {
+      if (ev.sourceType !== ScriptEventSource.Server) return;
+      handler(ev.id, ev.message);
+    };
+    system.afterEvents.scriptEventReceive.subscribe(listener);
+    return {
+      dispose: () => system.afterEvents.scriptEventReceive.unsubscribe(listener)
+    };
+  }
+  onReady(handler) {
+    const listener = (_ev) => handler();
+    world.afterEvents.worldLoad.subscribe(listener);
+    return {
+      dispose: () => world.afterEvents.worldLoad.unsubscribe(listener)
+    };
+  }
+  createIdRegistry(objectiveId) {
+    return new ScoreboardIdRegistry(objectiveId);
+  }
+  createRandom() {
+    return new SeedRandom(this.options.randomSeed);
+  }
+  bindEvents(handler) {
+    const disposables = [];
+    for (const [name, fn] of Object.entries(minecraftEventBinding.after)) {
+      disposables.push(
+        fn(world, (payload) => {
+          handler({ phase: "after", name, payload });
+        })
+      );
+    }
+    for (const [name, fn] of Object.entries(minecraftEventBinding.before)) {
+      disposables.push(
+        fn(world, (payload) => {
+          handler({ phase: "before", name, payload });
+        })
+      );
+    }
+    return {
+      dispose: () => disposables.forEach((d) => d.dispose())
+    };
+  }
+  scheduler = {
+    runInterval: (cb, tick) => system.runInterval(cb, tick),
+    runTimeout: (cb, tick) => system.runTimeout(cb, tick),
+    clearRun: (id) => system.clearRun(id)
+  };
+};
+
 // src/router/KairoRouter.ts
 var KairoRouter = class {
   kairoContext;
@@ -14844,12 +14813,11 @@ var KairoRouter = class {
     this.assertRunnable();
     this.scheduler.clearRun(runId);
   }
-  init(properties, options) {
+  init(properties) {
     if (this.kairoContext) {
       throw new KairoRouterInitError("AlreadyInitialized" /* AlreadyInitialized */);
     }
-    const runtimeOption = options?.runtime ?? "minecraft";
-    this.runtime = resolveRuntime(runtimeOption);
+    this.runtime = new KairoRuntime();
     this.scheduler = new KairoScheduler(this.runtime.scheduler);
     const { context, mutator } = createKairoContext(properties);
     this.kairoContext = context;
@@ -14965,12 +14933,6 @@ var KairoRouter = class {
     }
   }
 };
-function resolveRuntime(option) {
-  if (option === "minecraft") {
-    return new MinecraftRuntime();
-  }
-  return option;
-}
 function resolveRandom(runtime) {
   if ("createRandom" in runtime && typeof runtime.createRandom === "function") {
     return runtime.createRandom();
@@ -14997,13 +14959,14 @@ var MinecraftModule = /* @__PURE__ */ ((MinecraftModule2) => {
 var router = new KairoRouter();
 router.afterEvents.addonActivate.subscribe(() => {
 });
-world3.afterEvents.worldLoad.subscribe(() => {
+world2.afterEvents.worldLoad.subscribe(() => {
 });
 export {
   AddonActivateAfterEvent,
   AddonDeactivateBeforeEvent,
   KairoContext,
   KairoRouter,
+  KairoRuntime,
   MinecraftModule,
   SupportedTag,
   router
